middleware manual (funcao) para o codigo e verificação dos campos passados em create e edit

+++ API/library/views.py
@@ -1,7 +1,18 @@
 from django.shortcuts import render
 from django.http import JsonResponse
+from django.utils import timezone
 from library.models import Book
 import json
+    
+
+def verifyCode(code):
+    value = list(Book.objects.filter(code = code).values())
+
+    if(value == []):
+        return False
+
+    return True
+
 
 
 def bookList(request):
@@ -14,13 +25,23 @@ def bookCreate(request):
     if request.method == "POST":
         body_unicode = request.body.decode('utf-8')
         body = json.loads(body_unicode)
+
+        fields = ['code', 'title', 'author', 'date_launch', 'amount_available']
+
+        for field in fields:
+            try:
+                body[field]
+            except:
+                return JsonResponse({"erro":'campo ' + field + ' não fornecido'}, safe=False)
+        
+
         
         book = Book(
             code = body['code'],
             title = body['title'],
             author = body['author'],
             date_launch = body['date_launch'],
-            date_register = body['date_register'],
+            date_register = body['date_register'] if 'date_register' in body else timezone.now(),
             amount_available = body['amount_available']
         )
     
@@ -31,19 +52,15 @@ def bookCreate(request):
 
 
 def bookEdit(request, code):
-    if request.method == "PUT":
+    if request.method == "PUT" and verifyCode(code):
         body_unicode = request.body.decode('utf-8')
         body = json.loads(body_unicode)
 
         book = Book.objects.get(pk=code)
 
-        book.code = body['code'] if 'code' in body else book.code
-        book.title = body['title'] if 'title' in body else book.title
-        book.author = body['author'] if 'author' in body else book.author
-        book.date_launch = body['date_launch'] if 'date_launch' in body else book.date_launch
-        book.date_register = body['date_register'] if 'date_register' in body else book.date_register
-        book.amount_available = body['amount_available'] if 'amount_available' in body else book.amount_available
-    
+        for field in body:
+            setattr(book, field, body[field])
+            
         book.save()
 
     return JsonResponse({})
@@ -51,10 +68,10 @@ def bookEdit(request, code):
 
 
 def bookDelete(request, code):
-    if request.method == "DELETE":
+    if request.method == "DELETE" and verifyCode(code):
         book = Book.objects.get(code=code)
         book.delete()
-    
+
     return JsonResponse({})
